feat(msp): move time select to common
