+++ shell/app/common/components/time-select/time-select.scss
